requirements.txt
# Для роботи з великими мововими моделями через OpenRouter
openai>=1.0.0
# Для роботи з Git
GitPython>=3.1.0
# Для завантаження змінних оточення
python-dotenv>=1.0.0
# Для веб-інтерфейсу (опціонально)
# gradio>=4.0.0
# Для HTTP-запитів
requests>=2.30.0
# Для веб-пошуку
serpapi>=0.1.5